requirements.txt
# Zaktualizowana lista wymagań z Flask w wersji 2.2.5
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.0
google-generativeai>=0.5.0
Flask-Cors>=4.0.0
Flask-Bcrypt
Flask-JWT-Extended
gunicorn
